Revert "asd"
This reverts commit d0ee499ed740db93d6e0e9dfdce251b0e1aadb9e.

+++ Arquivo.h
@@ -18,7 +18,7 @@ typedef struct Letra{
 
 FILE *arq;
 unsigned char *vetorDeLetras;
-long int tamanho;
+int tamanho;
 Tabela *codigos;
 int quantosBytes;
 extern void lerArq(char *nome, char tipo);
